AP-5335: Remove disregarded benefits content from CFE result and means report for truelayer
We are no longer giving providers the oppertunity to add
disregarded beenfits (except housing benefits which are
are handled differently). So this content will always be zero going
forward and BA says could lead to confusion by case workers if they
find bank transactions that should not have been added/categorised
as benefits, for example.