AP-5317: Use default prometheus export sidekiq Instrumentation
An proemtheus custom alert rule was added that relies on
ruby_sidekiq_stats_dead_size stats. These do not appear to
be being generated and therefore no alerts triggered.

Changing the proemtheus exporter death handler and other Instrumentation
in line with those using in laa-submit-crime-forms in the hope
this will generate the necessary stats to alert on.