Disable digest extract, export and reset on uat
It just causes a lot of failed pods which eat up quotas.

This is to prevent the regular failures that branches
on UAT get which create pods that eat up our pod quota
and distract devs unnecessarily.

We can remove if we are want to actually change and test
on a branch.