AP-5416: Fix bug that keeps bank transactions when type has been removed
It was possible on the truelayer journey to remove an income or outgoing
type after bank transactions had been categorised as that type and the
categorisation would remain.

This does not make logic sense and, to follow the cash transaction example,
should instead removed the categorisation of any any bank transactions
categorised as that type.