AP-5143: initiate state machine after application creation
With the removal of this hook an inadvertent side effect has been
removed
```
  after_create do
    ActiveSupport::Notifications.instrument "dashboard.application_created", id:, state:
  end
````

The calling of the `state` method in this code means it actually calls
`state_machine_proxy` method, which creates a `state_machine` relation
on the application if it does not have one. Its removal therefore opens
up issues both in tests andpossibly functioning of the app.

By replacing it with an explcit call to `state_machine_proxy` aliased
as `find_or_create_state_machine` we avoid the issue.