AP-2910: Save attachment categorisation on select
Categorisation would get lost if any already uploaded
files were deleted or a new file uploaded. This was due
to the page being refreshed by the javascript when a
new file is uploaded or existing upload deleted.

This solution implements an immediate update of the attachment's
type/category when a categorisation is selected from the select-list.