AP-5524 extension: handle invalid format request with 404
Scripted probing attacks may use format modifiers on their
request. This handles them with a 404. The advatage
of this to not give the attacker info on what is a server
error and to not raise this an actual error in the code base.

We have an alert for 404s (exceeding 100) in a 24 hour period
which will alert us to a serious probing attack in any event.