AP-5566: Prevent infinite proceeding loop
When the provider says they do not wish to accept the
substantive default level of service then the Defaults
are cleared and they are taken to the page to select
non default levels of service. However the subsequent
page about levels of service does not update the `stage`,
leaving it nil.

On ending the proceeding loop for this proceeding it is checked
for completeness, which it fails because it has no level of service
details or they are incomplete due to the `stage` value not being
present.

This fix simply updates `stage` as well as `name`, thereby making
the substantive levels of service complete. The same has been applied
to the emegency levels of service form.