AP-5530: Add original (case) judge level question
Inline with designs this adds the second question for
the "second appeal", or "ECCT questions", flow to the
the merits questions.

NOTE: Using form based callbacks.
Agreed by developers to introduce a new pattern that clears the dependant
attributes, `appeal#original_judge_level` attribute in this case,
via a callback defined in the base form and set in any form that
wants to use it as opposed as an alterntive to both overriding the `save`
method or using similar active record model callbacks.