AP-4439: Display flash message when transaction removed
For accessibility we need to announce the removal of the
record to the page/screen reader. For progressive web app
reasons we need to ensure removal is also available and
announced to screen readers when javascript is disabled.

This provides mechanism that should be both accessible and
progressive. Accesible because it simply refreshes the page
thereby making use of the flash mechanism employed to
notify screen readers and progressive because it does the same
with javascript disabled.