AP-5647: Update proceeding captions
Swicth from proceedings.last to proceedings.order(:created_at).last

Because we use GUIDs as the id, there is no value in a plain .last
call as it orders by GUID text.  Adding the order allows us to get
the most recently created proceeding, which is what is expected,
without having to make the URLs and routing more complex by adding
specific proceeding IDs into the routes.

This also matches the existing use in the ProceedingIssueStatusesController