AP-5725: Update state change for linked cases
The method of changing the state from initiated to entering_applicant_details was only
being called on certain addresse and the limitation controllers.  When a provider chose
a business address and linked a case, it bypassed all of the state change mechanisms.

This PR moves the address state change to the firsy address page which should not be
bypassed in any of our current flows