AP-5669: Update check benefits page routing
Update to rails 8.0.1 flagged issue in routes where the check benefits
page was not routing correctly. This was due to the page using an index
action which is not allowed in singleton routes. This commit updates the
action to be a show action and updates the routes, removes a duplicate
route and updates any affected specs.