AP-5811: Spike Task list
Spike using NSCC task list but modified and then customised
to handle validating all forms and or models via the notion
of a "task validator".

This adds an editable applications state machine. It inherits from
the base_state_machine but overrides its `aasm` block, thereby
not implementing ANY states and events. This may allow us to build up the
state machine in isolation and transfer only what is needed.

It allows the CYA flow to continue as is, but if the users returns to the task
list then its behavior reverts to the primary/first-pass flow. This kind of what
we want, but designers may want the task list flow to be go back to the task list
between tasks when editing completed tasks.