AP-6089: Build restrictive tasklist pattern
Implement a pattern in the applcation code base that can be extended
to build the restrictive task list.

Based on a spike that used the NSCC task list but modified
and then customised to handle
- our own task list design customisations
  - sections with content-only under certain circumstances
  - sub-sections
  - subsections with content-only under certain circumstances
  - conditional task list items
- an example of determing a task status through validators, forms and or
  models via the notion of a "task validator".

This is non-navigable and is merely intended as a step to
build the task list.