AP-5073: Ensure statement of case error message has something to link
No field level error appears because the original_file dropzone field is hidden since
it is handled by JS. Without JS the error displays as expected.

This handle the situation by attaching the error to the always visible
statement text field as well as the original_file field. With JS enabled
or disabled you get a working link to the text field.