AP-5776: update the 'should be on' feature step
This was causing a race condition where the landed page would still
be in the middle of rendering when the url was checked.  Switching
to the have_current_path helper means that cucumber will wait for
the page render to complete before checking the url.

The prvious version would be checking the url _before_ the page had
switched and therefore fail... meanwhile the page was rendering and
the screenshot would show the expected page!