AP-5776 Fix bypassing evidence validation bug (#7705)
* AP-5776 Fix bypassing evidence validation bug

* Redirect to supporting evidence page when the `copy_of_court_order` or `child_care_assessment` answer is changed from 'No' to 'Yes' from the CYA merits page, so that the user has the opportunity to upload evidence for each of these. Update features and specs. 

* Add csrf enabled to feature tests - this enables allow_forgery_protection on the cucumber pages and
allows JS postbacks to work as expected - this is generally added
as needed rather than globally to prevent slow setup steps getting
repeated. I have also removed the 2 second sleep as there is no need, the page
works properly with csrf enabled

* Update the 'should be on' feature step - this was causing a race condition where the landed page would still
be in the middle of rendering when the url was checked.  Switching
to the have_current_path helper means that cucumber will wait for
the page render to complete before checking the url. The prvious version would be 
checking the url _before_ the page had switched and therefore fail... meanwhile the 
page was rendering and the screenshot would show the expected page!

---------

Co-authored-by: ColinBruce <colin.bruce@digital.justice.gov.uk>