AP-5646: Update two year digit handling
This was written a long time ago, when only date of birth was entered

We now have multiple date inputs that require future dates to be
entered, e.g. hearing dates in the future, and this handler would
reset a future date to be 19xx.

This update only applies a rollback on dates when the requested
field is a date of birth entry. For all others it will roll forward