Fix: Broken feature test
This started failing today but was fine yesterday. The cause seems to be sequencing, a
recent update re-recorded the VCR cassette expecting the first radio button in
the SE003 results to be the enforcement option.  The first option, today, is
the appeal option.  I have changed the radio button select to the explict
`Prohibited steps order - enforcement` option so that the VCR condition will be
met