AP-6008: Check flash obj rather than page content
Flash partial is now only displayed on pages which use a page template.
As this is a request spec, it redirects to the home page, which doesn't
use a page template. It didn't make sense to add the flash partial to
this page just for the benefit of a test passing, so the test has been
rewritten to check against the flash object instead.