AP-5519: Ensure consistent currency validation
Use the CurrencyCleaner class in the forms and handlers that
had a different solution used.  This ensures the same
substitutions are used in all forms and a consistent return
is provided by the system.

This is, possibly, a short term solution as we are using
many different ways of validating, storing and displaying
currency values between different forms and handlers