Fix: correct `app:` label for worker
This was changed in this PR for a particular reason but I am not
sure it is correct.

https://github.com/ministryofjustice/laa-apply-for-legal-aid/pull/5345

One impact of having a different `app:` label for worker pods, i believe,
is that the monitoring `netpol` will be applied to them, thus preventing
monitoring from the cloud platform hosted monitoring tooling, prometheus.

The `netpol` only applies to pods matching its selector, which is `app: {{ .Chart.Name }}``
which is `app=apply-for-legal-aid` ( `.Chart.Name` equals the Chart.yml name attribute).
this matches the `"apply-for-legal-aid.name"` value used as the `app:` label for metrics,
clamav and web pods, as defined in their respective deployments. Because the worker
deployment uses `app: {{ template "apply-for-legal-aid.name" . }}-worker` for its label it
will never be matched to the `netpol` which means monitoring tooling cannot "see" the worker.

I believe this is why the prometheus alerting that I put in place for sidekiq dead job
queue monitoring has never actually worked - i.e. despite the prometheus stats existing
in the worker worker pods they cannot be scraped by the monitoring tooling.

A note on labels:

`spec.selector.matchLabels`` is used to identify which pods/resources the deployment
controls.

`spec.template.metadata.labels` defines the labels that will be added to each pod/resource
spun-up by the deployment.

Together the above two sections control what pods are created or replaced and what their labels will
be. see this [medium article](https://medium.com/@zwhitchcox/matchlabels-labels-and-selectors-explained-in-detail-for-beginners-d421bdd05362) for more.

When changing matchLabels you will need to destroy the deployment entirely since
they are immutable and will raise an error to that effect.