Fix: Amend dump script inline with container changes
The separation of the clamav container into its own pod
requires changes to script to enable it to find a pod to use
for dump file. Namely, only 1 container in the pod and it does
not need the `-c web` therefore.

I have moved to using work pods specifically too. This is to avoid
overloading web server containers and avoids the potential to inadvertently
choose a clamav or metrics pod for the dum (which would then fail).