AP-6203: Change name of silas id column
Change from `silas_uiud` to `silas_id` for more idiomatic naming
and because uuid is a data type rather than a name for a column.

Note, we are keeping the silas_id as a string rather than making it
a datatype of uuid because we do not have confidence that SiLAS will
not change its claim enrocihment for USER_NAM to be something else in
the future.