AP-6215: Handle CYA flow issues when overriding DWP
When provider believes DWP results is wrong and provider is checking them
they should check both name, DoB and NI (separate page). This is an exception
to the normal CYA flow which wa previously implemented and locked down
by a feature test. This fixes the feature test reintroducing the behaviour for
the new flow.