AP-6215: Remove question over back hisotry
I don't think we need to delete the equivalent new page, dwp-results,
as I cannot see a reason not to allow the provider to go back to that
result page to perhaps choose "This is not correct" or "Yes, continue"
for negative scenario. For positive BC result scenarios it would not
provide any other option than Save and continue and similarly for DWP
down scenarios.

The BC check on that page is now effectively "cached"/stored so will not
be requeried aside from in the DWP down scenario.