Fix: get snyk scanning working again
This fixes and extends the snyk vulnerability scanning we use. This
tests code, dependencies and the container image against known
vulnerabilites in snyk's database/system.

Previously it was breaking with the following error:

```sh
Error: Code Scanning could not process the submitted SARIF file:
A delivery cannot contain multiple runs with the same category. Please upload a single run per category. For more information, see https://github.blog/changelog/2025-07-21-code-scanning-will-stop-combining-multiple-sarif-runs-uploaded-in-the-same-sarif-file/
```