Chore: suspect CCMS downtime scheduler cronjobs
CCMS is not having any scheduled downtime at present, so these cronjobs
are not needed. In addition suspending them builds a backlog of jobs
that then run all at once when they are re-enabled, which is not ideal and
jobs can fail as a result race conditions and/or pod resource limits.