AP-6435: Return parsed datastore application ID from submission
This is to return the application ID assigned by the datastore on submission,
which is currently provided in the Location header of the response. The previous expectation
of a response body has been removed as it is not currently provided and JSON standards consider
Location header a suitable if not more idiomatic way of providing the id of a created record.