Chore: Make bank holidays out-of-hours
This adds a config setting to allow an array of bank holidays to be read purely for
out-of-hours checks.  This replaces the current set up of calling the bank holiday
worker multiple times per page view.

It sets the test variables to the past and then uses that date in the new test, this
should prevent us being unable to run a test suite on an upcoming bank holiday if there
is an emergency!

The work to refactor the Bank Holiday handling will be done in a separate ticket, this gives
us three months to complete it before easter... if bank holiday shut downs are still a thing
at easter 2026?!