AP-6499: Move job enqueuing from BankHoliday to cronjob
Instead of relying on the bank holiday update worker jobs being enqueued by
calls to BankHoliday model we rely on a scheduled cronjob instead. This means fewer
calls to the API.

This change, however, does require a change to the job to create an initial bank holiday record
if none exists, otherwise the cronjob will fail. This is especially needed for UAT branches
that have no bank holiday records on them. This leverages the pre-existing
behaviour in the before_validation hook of the model to populate the dates from
the API when a new record is created, and then exits early from the worker as
the data is necessarily current.