Chore: Remove unrequired metadata labels from cronjob templates
I believe this PR shows that that metat data in question was added
originally to the admin-report cronjob, but was not necessary for the
functionality. It was then then it looks like copy pasted to the others.
Terefore, removing similar labels from other cronjob templates as they
are not required and serve no purpose that i can see.