ap-6398: Remove validation from matter_opposition model
This was causing cloning an application with a matter opposition to fail if there was no reason.
It appears not all matter_opposition (e.g. where the matter_opposition if for a PLF application) will have a matter opposition.
We still have validation for reason on the matter_opposed_reason_form.